19: Moving components to seprate folders
