+++ src/components/typography/SecondaryHeading/index.tsx
@@ -1,0 +1,11 @@
+import * as S from './style';
+
+interface Props {
+  level?: 1 | 2 | 3 | 4 | 5 | 6;
+  as?: string;
+}
+
+export const SecondaryHeading: React.FC<Props> = ({ level, as, ...otherProps }) => {
+  as = as || 'h2';
+  return <S.SecondaryHeading as={as} {...otherProps} />;
+};
