+++ src/components/typography/Text/style.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Paragraph = styled.p`
+  color: #111111;
+  font: normal normal 16px/1.7em 'Open Sans';
+`;
